# storage/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel",
        "description": "Documentation for the Application API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Laravel Server #1"
        }
    ],
    "paths": {
        "/user": {
            "get": {
                "summary": "",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "user"
                ]
            }
        },
        "/users/all": {
            "get": {
                "summary": "Display a listing of the resource.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "users"
                ]
            }
        },
        "/users/{id}": {
            "get": {
                "summary": "",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "tags": [
                    "users"
                ]
            }
        },
        "/users": {
            "post": {
                "summary": "Store a newly created resource in storage.",
                "description": "",
                "deprecated": false,
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "(Unauthorized) Invalid or missing Access Token"
                    }
                },
                "tags": [
                    "users"
                ]
            }
        }
    },
    "tags": [],
    "components": {
        "schemas": {
            "Training": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "description": "bigint",
                        "nullable": false,
                        "example": 1990971109
                    },
                    "name": {
                        "type": "string",
                        "description": "string(255)",
                        "nullable": false,
                        "example": "string"
                    },
                    "length": {
                        "type": "integer",
                        "description": "integer",
                        "nullable": false,
                        "example": 1437864821
                    },
                    "max_people": {
                        "type": "integer",
                        "description": "integer",
                        "nullable": false,
                        "example": 1922002395
                    },
                    "type": {
                        "type": "string",
                        "description": "string",
                        "nullable": false,
                        "example": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "datetime",
                        "nullable": true,
                        "example": "string"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "datetime",
                        "nullable": true,
                        "example": "string"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "length",
                    "max_people",
                    "type"
                ]
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "description": "bigint",
                        "nullable": false,
                        "example": 1124105973
                    },
                    "name": {
                        "type": "string",
                        "description": "string(255)",
                        "nullable": false,
                        "example": "string"
                    },
                    "email": {
                        "type": "string",
                        "description": "string(255)",
                        "nullable": false,
                        "example": "string"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "datetime",
                        "nullable": true,
                        "example": "string"
                    },
                    "two_factor_secret": {
                        "type": "string",
                        "description": "text(65535)",
                        "nullable": true,
                        "example": "string"
                    },
                    "two_factor_recovery_codes": {
                        "type": "string",
                        "description": "text(65535)",
                        "nullable": true,
                        "example": "string"
                    },
                    "two_factor_confirmed_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "datetime",
                        "nullable": true,
                        "example": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "datetime",
                        "nullable": true,
                        "example": "string"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "description": "datetime",
                        "nullable": true,
                        "example": "string"
                    },
                    "role": {
                        "type": "string",
                        "description": "string",
                        "default": "user",
                        "nullable": false,
                        "example": "string"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "email",
                    "role"
                ]
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}